# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: frp UI Backend API
  version: 0.1.0
servers:
  - url: http://149.118.158.112:18080
    description: Remote room control server
  - url: http://127.0.0.1:8080
    description: Local development control server
  - url: http://127.0.0.1:7410
    description: Windows local client backend
paths:
  /v1/groups:
    post:
      operationId: createGroup
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /v1/groups/{groupId}/join:
    post:
      operationId: joinGroup
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/groupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGroupRequest'
      responses:
        '200':
          description: Device token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinGroupResponse'
  /v1/rooms:
    get:
      operationId: listRooms
      tags: [control]
      responses:
        '200':
          description: XTCP rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomView'
    post:
      operationId: createRoom
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: XTCP room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoomResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/roomId'
      responses:
        '200':
          description: XTCP room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomView'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateRoom
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/roomId'
        - $ref: '#/components/parameters/xRoomId'
        - $ref: '#/components/parameters/xRoomDeviceId'
        - $ref: '#/components/parameters/xRoomDeviceToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: XTCP room updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomView'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Room device auth failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteRoom
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/roomId'
        - $ref: '#/components/parameters/xRoomId'
        - $ref: '#/components/parameters/xRoomDeviceId'
        - $ref: '#/components/parameters/xRoomDeviceToken'
      responses:
        '204':
          description: XTCP room deleted
        '401':
          description: Room device auth failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/join:
    post:
      operationId: joinRoom
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomRequest'
      responses:
        '200':
          description: XTCP room joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRoomResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid room code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/nodes:
    get:
      operationId: listNodes
      tags: [control]
      responses:
        '200':
          description: Nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Node'
    post:
      operationId: createNode
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
      responses:
        '201':
          description: Node created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
  /v1/nodes/{nodeId}:
    delete:
      operationId: deleteNode
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/nodeId'
      responses:
        '204':
          description: Node deleted
  /v1/exposures:
    get:
      operationId: listExposures
      tags: [control]
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Exposures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exposure'
  /v1/exposures/private:
    post:
      operationId: createPrivateExposure
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrivateExposureRequest'
      responses:
        '201':
          description: Private exposure created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposure'
  /v1/exposures/public:
    post:
      operationId: createPublicExposure
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublicExposureRequest'
      responses:
        '201':
          description: Public exposure created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposure'
  /v1/exposures/{exposureId}:
    patch:
      operationId: updateExposure
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/exposureId'
        - $ref: '#/components/parameters/xGroupId'
        - $ref: '#/components/parameters/xDeviceId'
        - $ref: '#/components/parameters/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExposureRequest'
      responses:
        '200':
          description: Exposure updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposure'
    delete:
      operationId: deleteExposure
      tags: [control]
      parameters:
        - $ref: '#/components/parameters/exposureId'
        - $ref: '#/components/parameters/xGroupId'
        - $ref: '#/components/parameters/xDeviceId'
        - $ref: '#/components/parameters/authorization'
      responses:
        '204':
          description: Exposure deleted
  /v1/access-routes:
    post:
      operationId: createAccessRoute
      tags: [control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessRouteRequest'
      responses:
        '201':
          description: Access route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRoute'
  /internal/frps/plugin:
    post:
      operationId: frpsPlugin
      tags: [control-internal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: frps plugin decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginDecision'
...
components:
  parameters:
    groupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
    roomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
    nodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: string
    exposureId:
      name: exposureId
      in: path
      required: true
      schema:
        type: string
...
    xGroupId:
      name: X-Group-ID
      in: header
      required: true
      schema:
        type: string
    xDeviceId:
      name: X-Device-ID
      in: header
      required: true
      schema:
        type: string
    authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
      description: Bearer device token
    xRoomId:
      name: X-Room-ID
      in: header
      required: true
      schema:
        type: string
    xRoomDeviceId:
      name: X-Room-Device-ID
      in: header
      required: true
      schema:
        type: string
    xRoomDeviceToken:
      name: X-Room-Device-Token
      in: header
      required: true
      schema:
        type: string
      description: Room device token. Authorization bearer token is also accepted by the server when this header is omitted.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CreateGroupRequest:
      type: object
      required: [groupId, password]
      properties:
        groupId:
          type: string
        password:
          type: string
          minLength: 8
    JoinGroupRequest:
      type: object
      required: [password, deviceName]
      properties:
        password:
          type: string
        deviceName:
          type: string
    JoinGroupResponse:
      type: object
      required: [groupId, deviceId, deviceToken]
      properties:
        groupId:
          type: string
        deviceId:
          type: string
        deviceToken:
          type: string
    Group:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateRoomRequest:
      type: object
      required: [name, deviceName]
      properties:
        name:
          type: string
        deviceName:
          type: string
    RoomView:
      type: object
      required: [id, name, serverName, frpsAddr, frpsPort, enabled, memberCount]
      properties:
        id:
          type: string
        name:
          type: string
        serverName:
          type: string
        hostDeviceId:
          type: string
        frpsAddr:
          type: string
        frpsPort:
          type: integer
        enabled:
          type: boolean
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoomDeviceView:
      type: object
      required: [id, roomId, name, role, createdAt, lastSeenAt]
      properties:
        id:
          type: string
        roomId:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [host, visitor]
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    CreateRoomResponse:
      type: object
      required: [room, roomCode, device, deviceToken]
      properties:
        room:
          $ref: '#/components/schemas/RoomView'
        roomCode:
          type: string
        device:
          $ref: '#/components/schemas/RoomDeviceView'
        deviceToken:
          type: string
    JoinRoomRequest:
      type: object
      required: [roomCode, deviceName]
      properties:
        roomCode:
          type: string
        deviceName:
          type: string
    JoinRoomResponse:
      type: object
      required: [room, device, deviceToken]
      properties:
        room:
          $ref: '#/components/schemas/RoomView'
        device:
          $ref: '#/components/schemas/RoomDeviceView'
        deviceToken:
          type: string
    UpdateRoomRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    CreateNodeRequest:
      type: object
      required: [name, serverAddr, frpsPort]
      properties:
        name:
          type: string
        serverAddr:
          type: string
        frpsPort:
          type: integer
        authToken:
          type: string
          description: Optional frps auth token when the public frps requires one.
        allowPorts:
          type: array
          items:
            $ref: '#/components/schemas/PortRange'
        domainSuffixes:
          type: array
          items:
            type: string
    Node:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        serverAddr:
          type: string
        frpsPort:
          type: integer
        allowPorts:
          type: array
          items:
            $ref: '#/components/schemas/PortRange'
        domainSuffixes:
          type: array
          items:
            type: string
    PortRange:
      type: object
      required: [from, to]
      properties:
        from:
          type: integer
        to:
          type: integer
...
    DeviceAuth:
      type: object
      required: [groupId, deviceId, deviceToken]
      properties:
        groupId:
          type: string
        deviceId:
          type: string
        deviceToken:
          type: string
    CreatePrivateExposureRequest:
      allOf:
        - $ref: '#/components/schemas/DeviceAuth'
        - type: object
          required: [name, localIP, localPort]
          properties:
            name:
              type: string
            localIP:
              type: string
            localPort:
              type: integer
    CreatePublicExposureRequest:
      allOf:
        - $ref: '#/components/schemas/DeviceAuth'
        - type: object
          required: [nodeId, name, protocol, localIP, localPort]
          properties:
            nodeId:
              type: string
            name:
              type: string
            protocol:
              type: string
              enum: [tcp, udp, http, https]
            localIP:
              type: string
            localPort:
              type: integer
            remotePort:
              type: integer
            domain:
              type: string
    UpdateExposureRequest:
      type: object
      properties:
        enabled:
          type: boolean
    Exposure:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        deviceId:
          type: string
        nodeId:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [private, public]
        protocol:
          type: string
        localIP:
          type: string
        localPort:
          type: integer
        remotePort:
          type: integer
        domain:
          type: string
        enabled:
          type: boolean
    CreateAccessRouteRequest:
      allOf:
        - $ref: '#/components/schemas/DeviceAuth'
        - type: object
          required: [exposureId, bindPort]
          properties:
            exposureId:
              type: string
            bindAddr:
              type: string
            bindPort:
              type: integer
            fallbackBindPort:
              type: integer
    AccessRoute:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        deviceId:
          type: string
        exposureId:
          type: string
        bindAddr:
          type: string
        bindPort:
          type: integer
        fallbackBindPort:
          type: integer
...
    PluginDecision:
      type: object
      required: [reject]
      properties:
        reject:
          type: boolean
        reject_reason:
          type: string
        unchanged:
          type: boolean